requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect2==0.7.0
Django==4.1.4
django-crispy-forms==1.14.0
django-environ==0.9.0
django-model-utils==1.1.0
djangorestframework==3.14.0
fonttools==4.38.0
fpdf==1.7.2
html5lib==1.1
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
numpy==1.24.1
pdfkit==1.0.0
Pillow==9.3.0
prompt-toolkit==3.0.36
pycparser==2.21
pydyf==0.5.0
pyphen==0.13.2
pytz==2022.7
redis==4.4.2
reportlab==3.6.12
six==1.16.0
sqlparse==0.4.3
tinycss2==1.2.1
vine==5.0.0
wcwidth==0.2.6
weasyprint==57.2
webencodings==0.5.1
zopfli==0.2.2
